refactor: added promotion module
