+++ crates/bench/src/main.rs
@@ -5,6 +5,7 @@ mod io;
 mod manifest;
 mod markdown;
 mod profiler;
+mod promotion;
 mod regression;
 mod report;
 mod utils;
@@ -13,18 +14,16 @@ use crate::benchmark::macro_bench::run_macro_suite;
 use crate::benchmark::micro_bench::run_micro_suite;
 use crate::cli::{Args, BenchSuite, FailOn};
 use crate::io::{
-    load_release_index, load_report, merge_reports, resolve_baseline_paths, write_release_index,
-    write_report,
+    load_release_index, load_report, merge_reports, resolve_baseline_paths, write_report,
 };
 use crate::manifest::load_manifest;
 use crate::markdown::write_markdown;
+use crate::promotion::run_promotion;
 use crate::regression::{RegressionSeverity, compare_reports, format_issue, resolve_gate_config};
-use crate::report::{BenchReport, ReleaseBaselineEntry};
+use crate::report::BenchReport;
 use crate::utils::{suite_label, unix_now};
 
 use clap::Parser;
-use std::fs;
-use std::path::PathBuf;
 
 fn print_summary(report: &BenchReport) {
     println!(
@@ -43,74 +42,6 @@ fn print_summary(report: &BenchReport) {
             case.latency_p99_ns
         );
     }
-}
-
-fn run_promotion(args: &Args) -> Result<(), String> {
-    let release = args
-        .promote_release
-        .as_ref()
-        .ok_or_else(|| "internal error: promote_release missing".to_string())?;
-
-    let mut index = load_release_index(&args.baseline_index)?;
-
-    let release_dir = PathBuf::from("bench").join("baselines").join(release);
-    fs::create_dir_all(&release_dir).map_err(|err| {
-        format!(
-            "failed to create release baseline directory '{}': {err}",
-            release_dir.display()
-        )
-    })?;
-
-    if !args.promote_micro_report.exists() {
-        return Err(format!(
-            "micro report '{}' does not exist",
-            args.promote_micro_report.display()
-        ));
-    }
-    if !args.promote_macro_report.exists() {
-        return Err(format!(
-            "macro report '{}' does not exist",
-            args.promote_macro_report.display()
-        ));
-    }
-
-    let micro_dest = release_dir.join("micro.json");
-    let macro_dest = release_dir.join("macro.json");
-
-    fs::copy(&args.promote_micro_report, &micro_dest).map_err(|err| {
-        format!(
-            "failed to copy micro report '{}' -> '{}': {err}",
-            args.promote_micro_report.display(),
-            micro_dest.display()
-        )
-    })?;
-    fs::copy(&args.promote_macro_report, &macro_dest).map_err(|err| {
-        format!(
-            "failed to copy macro report '{}' -> '{}': {err}",
-            args.promote_macro_report.display(),
-            macro_dest.display()
-        )
-    })?;
-
-    let entry = ReleaseBaselineEntry {
-        micro: micro_dest.to_string_lossy().to_string(),
-        macro_report: macro_dest.to_string_lossy().to_string(),
-    };
-    index.releases.insert(release.clone(), entry);
-    if args.set_current_release {
-        index.current_release = release.clone();
-    }
-
-    write_release_index(&args.baseline_index, &index)?;
-
-    println!(
-        "Promoted release baseline '{}' (micro='{}', macro='{}')",
-        release,
-        micro_dest.display(),
-        macro_dest.display()
-    );
-
-    Ok(())
 }
 
 fn main() -> Result<(), String> {
